uv-resolver: make source=git more structured
This splits out the git fields into an inline table instead of smushing
them into a single URL.

I'm unsure about this change, because it seems like we are using the git
URLs in other user facing ways. For example, some of the snapshot diffs
look like this:

         Installed 2 packages in [TIME]
          - project==0.1.0 (from file://[TEMP_DIR]/)
          + project==0.1.0 (from file://[TEMP_DIR]/)
    -     + uv-public-pypackage==0.1.0 (from git+https://github.com/astral-test/uv-public-pypackage@0dacfd662c64cb4ceb16e6cf65a157a8b715b979?rev=0.0.1#0dacfd662c64cb4ceb16e6cf65a157a8b715b979)
    +     + uv-public-pypackage==0.1.0 (from git+https://github.com/astral-test/uv-public-pypackage@0dacfd662c64cb4ceb16e6cf65a157a8b715b979)

which seems non-ideal. It should be possible to make the lock file use
structured fields while keeping the URL for terser output as above, but
I'm not sure that it's worth it.

This builds on top of #4627